phonebook

+++ part2/phonebook/src/App.jsx
@@ -3,12 +3,14 @@ import Filter from './components/Filter'
 import PersonForm from './components/PersonForm'
 import Persons from './components/Persons'
 import personService from './services/personService'
+import Notification from './components/Notification'
 
 const App = () => {
   const [persons, setPersons] = useState([])
   const [newName, setNewName] = useState('')
   const [newNumber, setNewNumber] = useState('')
   const [search, setSearch] = useState('')
+  const [errorMessage, setErrorMessage] = useState(null)
 
   const handleSearchChange = (event) => {
     setSearch(event.target.value)
@@ -24,18 +26,28 @@ const App = () => {
 
   const handleDelete = (id) => {
     const person = persons.find(p => p.id === id)
-    if (window.confirm(`Delete ${person.name}?`)) {
+    if (person && window.confirm(`Delete ${person.name}?`)) {
       personService
         .deletePerson(id)
         .then(() => {
           setPersons(persons.filter(p => p.id !== id))
         })
+        .catch(error => {
+          setErrorMessage(
+            `Information of ${person.name} has already been removed from server`
+          )
+          setTimeout(() => {
+            setErrorMessage(null)
+          }, 3000)
+        })
     }
+    
   }
 
   return (
     <div>
-      <h2>Phonebook</h2>
+      <h1>Phonebook</h1>
+      <Notification message={errorMessage} />
       <Filter search={search} handleSearchChange={handleSearchChange} />
       <h3>Add a new</h3>
       <PersonForm 
@@ -45,6 +57,8 @@ const App = () => {
         setNewName={setNewName} 
         newNumber={newNumber} 
         setNewNumber={setNewNumber} 
+        setErrorMessage={setErrorMessage}
+        errorMessage={errorMessage}
       />
       <h3>Numbers</h3>
       <Persons persons={persons} search={search} handleDelete={handleDelete} />
